+++ crates/astria-core/src/connect/mod.rs
@@ -3,227 +3,3 @@ pub mod market_map;
 pub mod oracle;
 pub mod service;
 pub mod types;
-
-pub mod transaction {
-    pub mod v1 {
-        use indexmap::IndexMap;
-        use tendermint::{
-            abci::types::ExtendedCommitInfo,
-            block::Round,
-        };
-
-        use crate::{
-            connect::types::v2::{
-                CurrencyPair,
-                CurrencyPairError,
-                CurrencyPairId,
-            },
-            generated::{
-                astria::protocol::connect::v1 as raw,
-                astria_vendored,
-            },
-        };
-
-        impl From<astria_vendored::tendermint::abci::ExtendedCommitInfo>
-            for tendermint_proto::abci::ExtendedCommitInfo
-        {
-            fn from(value: astria_vendored::tendermint::abci::ExtendedCommitInfo) -> Self {
-                tendermint_proto::abci::ExtendedCommitInfo {
-                    round: value.round,
-                    votes: value.votes.into_iter().map(Into::into).collect(),
-                }
-            }
-        }
-
-        impl From<astria_vendored::tendermint::abci::ExtendedVoteInfo>
-            for tendermint_proto::abci::ExtendedVoteInfo
-        {
-            fn from(value: astria_vendored::tendermint::abci::ExtendedVoteInfo) -> Self {
-                tendermint_proto::abci::ExtendedVoteInfo {
-                    validator: value.validator.map(Into::into),
-                    vote_extension: value.vote_extension,
-                    extension_signature: value.extension_signature,
-                    block_id_flag: value.block_id_flag,
-                }
-            }
-        }
-
-        impl From<astria_vendored::tendermint::abci::Validator> for tendermint_proto::abci::Validator {
-            fn from(value: astria_vendored::tendermint::abci::Validator) -> Self {
-                tendermint_proto::abci::Validator {
-                    address: value.address,
-                    power: value.power,
-                }
-            }
-        }
-
-        impl From<tendermint_proto::abci::ExtendedCommitInfo>
-            for astria_vendored::tendermint::abci::ExtendedCommitInfo
-        {
-            fn from(value: tendermint_proto::abci::ExtendedCommitInfo) -> Self {
-                astria_vendored::tendermint::abci::ExtendedCommitInfo {
-                    round: value.round,
-                    votes: value.votes.into_iter().map(Into::into).collect(),
-                }
-            }
-        }
-
-        impl From<tendermint_proto::abci::ExtendedVoteInfo>
-            for astria_vendored::tendermint::abci::ExtendedVoteInfo
-        {
-            fn from(value: tendermint_proto::abci::ExtendedVoteInfo) -> Self {
-                astria_vendored::tendermint::abci::ExtendedVoteInfo {
-                    validator: value.validator.map(Into::into),
-                    vote_extension: value.vote_extension,
-                    extension_signature: value.extension_signature,
-                    block_id_flag: value.block_id_flag,
-                }
-            }
-        }
-
-        impl From<tendermint_proto::abci::Validator> for astria_vendored::tendermint::abci::Validator {
-            fn from(value: tendermint_proto::abci::Validator) -> Self {
-                astria_vendored::tendermint::abci::Validator {
-                    address: value.address,
-                    power: value.power,
-                }
-            }
-        }
-
-        #[derive(Debug, Clone)]
-        pub struct ExtendedCommitInfoWithCurrencyPairMapping {
-            pub extended_commit_info: ExtendedCommitInfo,
-            pub id_to_currency_pair: IndexMap<CurrencyPairId, CurrencyPair>,
-        }
-
-        impl ExtendedCommitInfoWithCurrencyPairMapping {
-            #[must_use]
-            pub fn new(
-                extended_commit_info: ExtendedCommitInfo,
-                id_to_currency_pair: IndexMap<CurrencyPairId, CurrencyPair>,
-            ) -> Self {
-                Self {
-                    extended_commit_info,
-                    id_to_currency_pair,
-                }
-            }
-
-            #[must_use]
-            pub fn empty(round: Round) -> Self {
-                Self {
-                    extended_commit_info: ExtendedCommitInfo {
-                        round,
-                        votes: Vec::new(),
-                    },
-                    id_to_currency_pair: IndexMap::new(),
-                }
-            }
-
-            #[must_use]
-            pub fn extended_commit_info(&self) -> &ExtendedCommitInfo {
-                &self.extended_commit_info
-            }
-
-            #[must_use]
-            pub fn id_to_currency_pair(&self) -> &IndexMap<CurrencyPairId, CurrencyPair> {
-                &self.id_to_currency_pair
-            }
-
-            pub fn try_from_raw(
-                raw: raw::ExtendedCommitInfoWithCurrencyPairMapping,
-            ) -> Result<Self, ExtendedCommitInfoWithCurrencyPairMappingError> {
-                let Some(extended_commit_info) = raw.extended_commit_info else {
-                    return Err(
-                        ExtendedCommitInfoWithCurrencyPairMappingError::field_not_set(
-                            "extended_commit_info",
-                        ),
-                    );
-                };
-                let extended_commit_info = ExtendedCommitInfo::try_from(
-                    <astria_vendored::tendermint::abci::ExtendedCommitInfo as Into<
-                        tendermint_proto::abci::ExtendedCommitInfo,
-                    >>::into(extended_commit_info),
-                )
-                .map_err(ExtendedCommitInfoWithCurrencyPairMappingError::extended_commit_info)?;
-                let id_to_currency_pair = raw
-                    .id_to_currency_pair
-                    .into_iter()
-                    .map(|id_with_currency_pair| {
-                        let currency_pair_id = CurrencyPairId::new(id_with_currency_pair.id);
-                        let Some(currency_pair) = id_with_currency_pair.currency_pair else {
-                            return Err(
-                                ExtendedCommitInfoWithCurrencyPairMappingError::field_not_set(
-                                    "currency_pair",
-                                ),
-                            );
-                        };
-                        let currency_pair = CurrencyPair::try_from_raw(currency_pair).map_err(
-                            ExtendedCommitInfoWithCurrencyPairMappingError::currency_pair,
-                        )?;
-                        Ok((currency_pair_id, currency_pair))
-                    })
-                    .collect::<Result<
-                        IndexMap<CurrencyPairId, CurrencyPair>,
-                        ExtendedCommitInfoWithCurrencyPairMappingError,
-                    >>()?;
-                Ok(Self {
-                    extended_commit_info,
-                    id_to_currency_pair,
-                })
-            }
-
-            pub fn into_raw(self) -> raw::ExtendedCommitInfoWithCurrencyPairMapping {
-                let extended_commit_info: tendermint_proto::abci::ExtendedCommitInfo =
-                    self.extended_commit_info.into();
-                let id_to_currency_pair = self
-                    .id_to_currency_pair
-                    .into_iter()
-                    .map(
-                        |(currency_pair_id, currency_pair)| raw::IdWithCurrencyPair {
-                            id: currency_pair_id.get(),
-                            currency_pair: Some(currency_pair.into_raw()),
-                        },
-                    )
-                    .collect();
-                raw::ExtendedCommitInfoWithCurrencyPairMapping {
-                    extended_commit_info: Some(extended_commit_info.into()),
-                    id_to_currency_pair,
-                }
-            }
-        }
-
-        #[derive(Debug, thiserror::Error)]
-        #[error(transparent)]
-        pub struct ExtendedCommitInfoWithCurrencyPairMappingError(
-            ExtendedCommitInfoWithCurrencyPairMappingErrorKind,
-        );
-
-        impl ExtendedCommitInfoWithCurrencyPairMappingError {
-            fn field_not_set(field: &'static str) -> Self {
-                Self(
-                    ExtendedCommitInfoWithCurrencyPairMappingErrorKind::FieldNotSet {
-                        field,
-                    },
-                )
-            }
-
-            fn currency_pair(error: CurrencyPairError) -> Self {
-                Self(ExtendedCommitInfoWithCurrencyPairMappingErrorKind::CurrencyPair(error))
-            }
-
-            fn extended_commit_info(error: tendermint::Error) -> Self {
-                Self(ExtendedCommitInfoWithCurrencyPairMappingErrorKind::ExtendedCommitInfo(error))
-            }
-        }
-
-        #[derive(Debug, thiserror::Error)]
-        enum ExtendedCommitInfoWithCurrencyPairMappingErrorKind {
-            #[error("field not set: {field}")]
-            FieldNotSet { field: &'static str },
-            #[error("invalid currency pair")]
-            CurrencyPair(#[from] CurrencyPairError),
-            #[error("invalid extended commit info")]
-            ExtendedCommitInfo(#[from] tendermint::Error),
-        }
-    }
-}
