+++ crates/astria-core/src/protocol/mod.rs
@@ -8,6 +8,7 @@ pub mod abci;
 pub mod account;
 pub mod asset;
 pub mod bridge;
+pub mod connect;
 pub mod fees;
 pub mod genesis;
 pub mod memos;
