+++ crates/astria-sequencer/src/app/test_utils.rs
@@ -186,7 +186,7 @@ pub(crate) fn transactions_with_extended_commit_info_and_commitments(
     txs: &[Transaction],
     deposits: Option<HashMap<RollupId, Vec<Deposit>>>,
 ) -> Vec<Bytes> {
-    use astria_core::connect::transaction::v1::ExtendedCommitInfoWithCurrencyPairMapping;
+    use astria_core::protocol::connect::v1::ExtendedCommitInfoWithCurrencyPairMapping;
     use prost::Message as _;
 
     use crate::proposal::commitment::generate_rollup_datas_commitment;
